# SPU LLVM: Optimize branching