Optimizes BRZ/BRNZ by using vpmovmskb and a check against the sign bit where possible.

For whatever reason vpextrd is 2 uops on both intel and ryzen cpus, while vpmovmskb is just 1 uop. 

It's only a small speedup but the optimization is pretty common (about 5,000/10,000 BRZ/BRNZ in wipeout HD can hit this optimization)