@Yahfz 2021-05-17
Master

SPU Integer Perf completed in 1368 ms (PS3: 8666 ms)
SPU Float Perf completed in 370 ms (PS3: 2379 ms)

SPU Integer Perf completed in 1365 ms (PS3: 8666 ms)
SPU Float Perf completed in 373 ms (PS3: 2379 ms)

PR

SPU Integer Perf completed in 1349 ms (PS3: 8666 ms)
SPU Float Perf completed in 368 ms (PS3: 2379 ms)

SPU Integer Perf completed in 1349 ms (PS3: 8666 ms)
SPU Float Perf completed in 367 ms (PS3: 2379 ms)