@Nekotekina 2021-05-17
Can do similar thing with BRHZ/BRHNZ I believe